# backend/Billullo.Api/swagger.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Billullo.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/"
    }
  ],
  "paths": {
...
    "/api/EmailParsingRules": {
      "get": {
        "tags": [
          "EmailParsingRules"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmailParsingRuleDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmailParsingRuleDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmailParsingRuleDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "EmailParsingRules"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmailParsingRuleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmailParsingRuleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmailParsingRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmailParsingRuleDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailParsingRuleDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailParsingRuleDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/EmailParsingRules/{id}": {
      "get": {
        "tags": [
          "EmailParsingRules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "string"
              ],
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmailParsingRuleDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailParsingRuleDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailParsingRuleDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "EmailParsingRules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "string"
              ],
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmailParsingRuleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmailParsingRuleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmailParsingRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmailParsingRuleDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailParsingRuleDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailParsingRuleDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "EmailParsingRules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "string"
              ],
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/EmailParsingRules/test": {
      "post": {
        "tags": [
          "EmailParsingRules"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestEmailParsingRuleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TestEmailParsingRuleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TestEmailParsingRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TestEmailParsingResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestEmailParsingResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestEmailParsingResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/Rules": {
      "get": {
        "tags": [
          "Rules"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryRuleDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryRuleDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryRuleDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Rules"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRuleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRuleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRuleDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRuleDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRuleDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Rules/{id}": {
      "get": {
        "tags": [
          "Rules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "string"
              ],
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRuleDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRuleDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRuleDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Rules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "string"
              ],
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryRuleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryRuleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRuleDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRuleDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRuleDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Rules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "string"
              ],
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Rules/match": {
      "get": {
        "tags": [
          "Rules"
        ],
        "parameters": [
          {
            "name": "description",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RuleMatchResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleMatchResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleMatchResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/Rules/by-category/{categoryId}": {
      "delete": {
        "tags": [
          "Rules"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "string"
              ],
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryRuleDto": {
        "required": [
          "id",
          "pattern",
          "categoryId",
          "categoryName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": [
              "integer",
              "string"
            ],
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$"
          },
          "pattern": {
            "type": "string"
          },
          "categoryId": {
            "type": [
              "integer",
              "string"
            ],
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$"
          },
          "categoryName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "CreateCategoryRuleRequest": {
        "required": [
          "pattern",
          "categoryId"
        ],
        "type": "object",
        "properties": {
          "pattern": {
            "type": "string"
          },
          "categoryId": {
            "type": [
              "integer",
              "string"
            ],
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$"
          }
        }
      },
      "CreateEmailParsingRuleRequest": {
        "required": [
          "name",
          "senderAddress",
          "subjectPattern",
          "amountRegex",
          "dateRegex",
          "dateFormat",
          "currencyFixed",
          "currencyRegex",
          "descriptionFixed",
          "descriptionRegex",
          "transactionType",
          "categoryId"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "senderAddress": {
            "type": [
              "null",
              "string"
            ]
          },
          "subjectPattern": {
            "type": [
              "null",
              "string"
            ]
          },
          "amountRegex": {
            "type": "string"
          },
          "dateRegex": {
            "type": "string"
          },
          "dateFormat": {
            "type": "string"
          },
          "currencyFixed": {
            "type": [
              "null",
              "string"
            ]
          },
          "currencyRegex": {
            "type": [
              "null",
              "string"
            ]
          },
          "descriptionFixed": {
            "type": [
              "null",
              "string"
            ]
          },
          "descriptionRegex": {
            "type": [
              "null",
              "string"
            ]
          },
          "transactionType": {
            "type": "string"
          },
          "categoryId": {
            "type": [
              "integer",
              "null",
              "string"
            ],
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$"
          },
          "priority": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32",
            "default": 0
          }
        }
      },
...
      "EmailParsingRuleDto": {
        "required": [
          "id",
          "name",
          "senderAddress",
          "subjectPattern",
          "amountRegex",
          "dateRegex",
          "dateFormat",
          "currencyFixed",
          "currencyRegex",
          "descriptionFixed",
          "descriptionRegex",
          "transactionType",
          "categoryId",
          "categoryName",
          "priority"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": [
              "integer",
              "string"
            ],
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$"
          },
          "name": {
            "type": "string"
          },
          "senderAddress": {
            "type": [
              "null",
              "string"
            ]
          },
          "subjectPattern": {
            "type": [
              "null",
              "string"
            ]
          },
          "amountRegex": {
            "type": "string"
          },
          "dateRegex": {
            "type": "string"
          },
          "dateFormat": {
            "type": "string"
          },
          "currencyFixed": {
            "type": [
              "null",
              "string"
            ]
          },
          "currencyRegex": {
            "type": [
              "null",
              "string"
            ]
          },
          "descriptionFixed": {
            "type": [
              "null",
              "string"
            ]
          },
          "descriptionRegex": {
            "type": [
              "null",
              "string"
            ]
          },
          "transactionType": {
            "type": "string"
          },
          "categoryId": {
            "type": [
              "integer",
              "null",
              "string"
            ],
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$"
          },
          "categoryName": {
            "type": [
              "null",
              "string"
            ]
          },
          "priority": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "RuleMatchResult": {
        "required": [
          "categoryId",
          "categoryName",
          "conflicts",
          "matches"
        ],
        "type": "object",
        "properties": {
          "categoryId": {
            "type": [
              "integer",
              "null",
              "string"
            ],
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$"
          },
          "categoryName": {
            "type": [
              "null",
              "string"
            ]
          },
          "conflicts": {
            "type": "boolean"
          },
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryRuleDto"
            }
          }
        }
      },
...
      "TestEmailParsingResult": {
        "required": [
          "matched",
          "amount",
          "date",
          "currency",
          "description",
          "error"
        ],
        "type": "object",
        "properties": {
          "matched": {
            "type": "boolean"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "date": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "error": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "TestEmailParsingRuleRequest": {
        "required": [
          "emailBody",
          "emailSubject",
          "senderAddress",
          "amountRegex",
          "dateRegex",
          "dateFormat",
          "currencyFixed",
          "currencyRegex",
          "descriptionFixed",
          "descriptionRegex"
        ],
        "type": "object",
        "properties": {
          "emailBody": {
            "type": "string"
          },
          "emailSubject": {
            "type": "string"
          },
          "senderAddress": {
            "type": "string"
          },
          "amountRegex": {
            "type": "string"
          },
          "dateRegex": {
            "type": "string"
          },
          "dateFormat": {
            "type": "string"
          },
          "currencyFixed": {
            "type": [
              "null",
              "string"
            ]
          },
          "currencyRegex": {
            "type": [
              "null",
              "string"
            ]
          },
          "descriptionFixed": {
            "type": [
              "null",
              "string"
            ]
          },
          "descriptionRegex": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UpdateCategoryRuleRequest": {
        "required": [
          "pattern",
          "categoryId"
        ],
        "type": "object",
        "properties": {
          "pattern": {
            "type": "string"
          },
          "categoryId": {
            "type": [
              "integer",
              "string"
            ],
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$"
          }
        }
      },
      "UpdateEmailParsingRuleRequest": {
        "required": [
          "name",
          "senderAddress",
          "subjectPattern",
          "amountRegex",
          "dateRegex",
          "dateFormat",
          "currencyFixed",
          "currencyRegex",
          "descriptionFixed",
          "descriptionRegex",
          "transactionType",
          "categoryId"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "senderAddress": {
            "type": [
              "null",
              "string"
            ]
          },
          "subjectPattern": {
            "type": [
              "null",
              "string"
            ]
          },
          "amountRegex": {
            "type": "string"
          },
          "dateRegex": {
            "type": "string"
          },
          "dateFormat": {
            "type": "string"
          },
          "currencyFixed": {
            "type": [
              "null",
              "string"
            ]
          },
          "currencyRegex": {
            "type": [
              "null",
              "string"
            ]
          },
          "descriptionFixed": {
            "type": [
              "null",
              "string"
            ]
          },
          "descriptionRegex": {
            "type": [
              "null",
              "string"
            ]
          },
          "transactionType": {
            "type": "string"
          },
          "categoryId": {
            "type": [
              "integer",
              "null",
              "string"
            ],
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$"
          },
          "priority": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32",
            "default": 0
          }
        }
      },
...
    }
  },
...
}